Least Privilege Edits

+++ app/Http/Controllers/Api/V1/AuthorTicketsController.php
@@ -9,7 +9,6 @@ use App\Http\Requests\Api\V1\UpdateTicketRequest;
 use App\Http\Resources\Api\V1\TicketResource;
 use App\Models\Ticket;
 use App\Policies\V1\TicketPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
  
 class AuthorTicketsController extends ApiController
@@ -22,30 +21,27 @@ class AuthorTicketsController extends ApiController
 
      public function store(StoreTicketRequest $request, $author_id)
     {
-        try {
-            $this->isAble('store', Ticket::class);
-
+        if ($this->isAble('store', Ticket::class)) {
             return new TicketResource(Ticket::create($request->mappedAttributes([
                 'author' => 'user_id'
             ]))); 
-        } catch (AuthorizationException $exception) {
-            return $this->error('You Are Not Authorized To Create This Ticket', 401);
         }
+
+        return $this->error('You Are Not Authorized To Create This Ticket', 401);
     }
 
     public function update(UpdateTicketRequest $request, $author_id, $ticket_id) {
         try {
             $ticket = Ticket::where('id', $ticket_id)->where('user_id', $author_id)->firstOrFail();
 
-            $this->isAble('update', $ticket);
+            if ($this->isAble('update', $ticket)) {
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            $ticket->update($request->mappedAttributes());
-            return new TicketResource($ticket);
-
+            return $this->error('You Are Not Authorized To Update This Ticket', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket Not Found', 404);
-        } catch (AuthorizationException $exception) {
-            return $this->error('You Are Not Authorized To Update This Ticket', 401);
         }
     }
 
@@ -53,15 +49,14 @@ class AuthorTicketsController extends ApiController
         try {
             $ticket = Ticket::where('id', $ticket_id)->where('user_id', $author_id)->firstOrFail();
 
-            $this->isAble('replace', $ticket);
+            if ($this->isAble('replace', $ticket)) {
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            $ticket->update($request->mappedAttributes());
-            return new TicketResource($ticket);
-
+            return $this->error('You Are Not Authorized To Update This Ticket', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket Not Found', 404);
-        } catch (AuthorizationException $exception) {
-            return $this->error('You Are Not Authorized To Update This Ticket', 401);
         }
     }
 
@@ -70,15 +65,14 @@ class AuthorTicketsController extends ApiController
         try {
             $ticket = Ticket::where('id', $ticket_id)->where('user_id', $author_id)->firstOrFail();
 
-            $this->isAble('destroy', $ticket);
+            if ($this->isAble('destroy', $ticket)) {
+                $ticket->delete();
+                return $this->ok('Ticket Successfuly Deleted');
+            }
 
-            $ticket->delete();
-            return $this->ok('Ticket Successfuly Deleted');
-                        
+            return $this->error('You Are Not Authorized To Delete This Ticket', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket Not Found', 404);
-        } catch (AuthorizationException $exception) {
-            return $this->error('You Are Not Authorized To Delete This Ticket', 401);
         }
     }
